Fix join_mask

+++ src/base_taint_analysis.h
@@ -7,6 +7,6 @@ struct Node {
     int first_transfer_index = -1;
     int predecessor_index[5] = { -1,-1,-1,-1,-1 };
     int successor_index[5] = { -1,-1,-1,-1,-1 };
-    BitVector join_mask = INT32_MAX;
+    BitVector join_mask = INT64_MAX;
 };
 }
